tls: move legacy code into own file
