+++ lib/internal/tls/secure-context.js
@@ -5,12 +5,8 @@ const {
   ArrayPrototypeFilter,
   ArrayPrototypeForEach,
   ArrayPrototypeJoin,
-  ArrayPrototypePush,
-  StringPrototypeIndexOf,
-  StringPrototypeSlice,
   StringPrototypeSplit,
   StringPrototypeStartsWith,
-  ObjectCreate,
 } = primordials;
 
 const {
@@ -41,28 +37,6 @@ const {
     TLS1_3_VERSION,
   },
 } = internalBinding('constants');
-
-// Example:
-// C=US\nST=CA\nL=SF\nO=Joyent\nOU=Node.js\nCN=ca1\nemailAddress=ry@clouds.org
-function parseCertString(s) {
-  const out = ObjectCreate(null);
-  ArrayPrototypeForEach(StringPrototypeSplit(s, '\n'), (part) => {
-    const sepIndex = StringPrototypeIndexOf(part, '=');
-    if (sepIndex > 0) {
-      const key = StringPrototypeSlice(part, 0, sepIndex);
-      const value = StringPrototypeSlice(part, sepIndex + 1);
-      if (key in out) {
-        if (!ArrayIsArray(out[key])) {
-          out[key] = [out[key]];
-        }
-        ArrayPrototypePush(out[key], value);
-      } else {
-        out[key] = value;
-      }
-    }
-  });
-  return out;
-}
 
 function getDefaultEcdhCurve() {
   // We do it this way because DEFAULT_ECDH_CURVE can be
@@ -340,5 +314,4 @@ function configSecureContext(context, options = {}, name = 'options') {
 
 module.exports = {
   configSecureContext,
-  parseCertString,
 };
